sdk progress

+++ include/revolution/ax/AXComp.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif
 
-extern u16 __AXCompressorTable[];
+extern u8 __AXCompressorTable[];
 
 #ifdef __cplusplus
 }
